requirements.txt
flask==2.0.3
werkzeug==2.0.3
python-dateutil==2.8.2
pytz==2021.3
pymongo==3.12.0
flask-pymongo==2.3.0
